Implement unauthenticated article serializer
Unauthenticated users will receive a limited data
set. For this reason, the article serializer for
unauthenticated users makes use of three custom
serializer method fields which return the article
owner' email (possibly to be amended with the
user's name at a later stage), the net votes
accumulated on the article, as well as the
number of comments.