Refactor article vote casting
A more nuanced logic has been implemented to allow
users changing their votes. The update method in
article_votes_views module has been amended to
check if the user has cast a vote before. If true,
the method checks if it was cast in the same
direction. If so, an error with message is
returned. If the new vote is cast into the other
direction the method now deletes the old vote and
replaces it with a new one.
If the user has not cast a vote before on
specified article then the same logic applies as
before.